Add extra overwrite flag for any scripts or raster/vector unpacking steps
A single command can have overwrite easily added to it, but scripts need to implement that, so -o is convenient support for overwrite without extra work for the user.

Unpacking step on the remote would always overwrite, now user can request overwrite with the same safety logic as for other commands.